[BUGFIX] Check if install tool enable file exists before removing it
The is_file() check got lost during the refactoring in
d855b4707ba1198b6c1dbf856c4377e230da4c4b.

Resolves: #80354
Releases: master, 7.6
Reviewed-on: https://review.typo3.org/52100
Tested-by: TYPO3com <no-reply@typo3.com>
Tested-by: Josef Glatz <josef.glatz@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>

+++ typo3/sysext/install/Classes/Controller/Action/Step/DefaultConfiguration.php
@@ -133,7 +133,7 @@ For each website you need a TypoScript template on the main page of your website
         );
         $formProtection->clean();
 
-        if (!EnableFileService::isInstallToolEnableFilePermanent()) {
+        if (EnableFileService::installToolEnableFileExists() && !EnableFileService::isInstallToolEnableFilePermanent()) {
             EnableFileService::removeInstallToolEnableFile();
         }
